+++ src/main/java/com/eEducation/ftn/repository/ColloquiumResultRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.ColloquiumResult;
 
-public interface ColloquiumResultRepository extends JpaRepository<ColloquiumResult, Long> {
+public interface ColloquiumResultRepository extends JpaRepository<ColloquiumResult, Integer> {
 
 }
