+++ src/main/java/com/eEducation/ftn/repository/CourseFileRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.CourseFile;
 
-public interface CourseFileRepository extends JpaRepository<CourseFile, Long> {
+public interface CourseFileRepository extends JpaRepository<CourseFile, Integer> {
 
 }
