+++ src/main/java/com/eEducation/ftn/repository/ExamPeriodRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.ExamPeriod;
 
-public interface ExamPeriodRepository extends JpaRepository<ExamPeriod, Long> {
+public interface ExamPeriodRepository extends JpaRepository<ExamPeriod, Integer> {
 
 }
