+++ src/main/java/com/eEducation/ftn/repository/GradeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.Grade;
 
-public interface GradeRepository extends JpaRepository<Grade, Long> {
+public interface GradeRepository extends JpaRepository<Grade, Integer> {
 
 }
