+++ src/main/java/com/eEducation/ftn/repository/RankRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.Rank;
 
-public interface RankRepository extends JpaRepository<Rank, Long>{
+public interface RankRepository extends JpaRepository<Rank, Integer>{
 
 }
